fix: restructure, offline refactors
