+++ src/index.js
@@ -228,10 +228,20 @@ import {
 } from './services/multi-user-accounts/multi-user-accounts.ts';
 
 import {
-	getPracticeSessionsOffline,
-	getRecentActivityOffline,
+	getRecentActivityOffline
+} from './services/offline/activities.ts';
+
+import {
+	getPracticeSessionsOffline
+} from './services/offline/practices.ts';
+
+import {
+	contentStatusCompletedManyOffline,
+	contentStatusCompletedOffline,
+	contentStatusResetOffline,
+	contentStatusStartedOffline,
 	recordWatchSessionOffline
-} from './services/offline/offline.ts';
+} from './services/offline/progress.ts';
 
 import {
 	emitProgressSaved,
@@ -489,8 +499,12 @@ export {
 	confirmEmailChange,
 	contentStatusCompleted,
 	contentStatusCompletedMany,
+	contentStatusCompletedManyOffline,
+	contentStatusCompletedOffline,
 	contentStatusReset,
+	contentStatusResetOffline,
 	contentStatusStarted,
+	contentStatusStartedOffline,
 	convertToTimeZone,
 	createAccount,
 	createComment,
